+++ src/components/EtGrid.vue
@@ -17,8 +17,12 @@
     </template>
 
     <!-- slot name syntax: body-cell-<column_name> -->
-    <q-td class="q-pa-none" auto-width slot="body-cell-edit" slot-scope="props" :props="props">
+    <q-td v-show="!disableEditting" class="q-pa-none" auto-width slot="body-cell-edit" slot-scope="props" :props="props">
       <q-btn round flat wait-for-ripple dense icon="settings" color="secondary" @click="setEditingRec(props.row)">
+      </q-btn>
+    </q-td>
+    <q-td v-show="!expandedCol.hidden" class="q-pa-none" auto-width :slot="getExpandedColName" slot-scope="props" :props="props">
+      <q-btn round flat wait-for-ripple dense :icon="expandedCol.icon" color="secondary" @click="excuteExpanedColAction(props.row,expandedCol.action)">
       </q-btn>
     </q-td>
 
@@ -46,6 +50,23 @@ export default {
     type: {
       default: 'xxx',
       type: String,
+    },
+    disableEditting: {
+      default: false,
+      type: Boolean,
+    },
+    expandedCol: {
+      default: () => {
+        return {
+          fieldName: 'id',
+          hidden: true,
+          icon: 'settings',
+          action: function(payload) {
+            console.log(payload)
+          },
+        }
+      },
+      type: Object,
     },
     hideSelection: Boolean,
     hideAdd: Boolean,
@@ -87,6 +108,9 @@ export default {
         this.$store.commit(this.type + '/setSelected', val)
       },
     },
+    getExpandedColName() {
+      return 'body-cell-' + this.expandedCol.fieldName
+    },
   },
   methods: {
     ...mapActions({
@@ -105,10 +129,12 @@ export default {
     selectedLabel(rowsNo) {
       return `Selected ${rowsNo}`
     },
+    excuteExpanedColAction(payload, action) {
+      action(payload)
+    },
   },
 }
 </script>
-
 <style scoped lang="stylus">
 .q-pa-none
   padding 0 !important