+++ src/store/giftcard/getters.js
@@ -1,6 +1,43 @@
 export const getIsLoading = state => {
   return state.isLoading
 }
-export const getCurrentGiftCardCode = state => {
-  return state.currentGiftCardCode
+
+export const getIsModalOpened = state => {
+  return state.isModalOpened
+}
+
+export const getIsModalQRCodeOpened = state => {
+  return state.isModalQRCodeOpened
+}
+
+export const getCurrentGenGiftCardCode = state => {
+  return state.currentGenGiftCardCode
+}
+
+export const getCustomer = state => {
+  return state.customer
+}
+
+export const getRecs = state => {
+  return state.recs
+}
+
+export const getCols = state => {
+  return state.cols
+}
+
+export const getFields = state => {
+  return state.fields
+}
+
+export const getTitle = state => {
+  return state.title
+}
+
+export const getIcon = state => {
+  return state.icon
+}
+
+export const getSelected = state => {
+  return state.selected
 }