+++ CHANGELOG.md
@@ -18,7 +18,7 @@ All notable changes to this project will be documented in this file.
 - Generate gift card QR
 - Topup page
 - Authenticate a GiftCard utility function
-- Apply gift card function
+- GiftCard page [@phillip](https://github.com/hemvanh/TPTEA/pull/55)
 ### Removed 
 - Join now button
 - Remove parameter "expiry" and change parameter "amount" to "giftCardid" in genGiftCard function